requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
requests==2.31.0

# Singapore Capstone Project Dependencies
pandas==2.1.4
numpy==1.24.3
scikit-learn==1.3.2
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0
streamlit==1.29.0
joblib==1.3.2

# Additional data processing
openpyxl==3.1.2
xlsxwriter==3.1.9

# Singapore-specific data sources
beautifulsoup4==4.12.2
lxml==4.9.3
